ml/pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.0"
pandas = "^2.0.2"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
scikit-learn = "^1.2.2"
lightgbm = "^3.3.5"
pydantic = "^1.10.9"
hydra-core = "^1.3.2"
mlflow = "^2.4.1"
pandera = "0.8.0"
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.4.1"
ipykernel = "^6.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
